+++ zetaclient/zetacore_observer.go
@@ -3,7 +3,6 @@ package zetaclient
 import (
 	"fmt"
 	"math"
-	"strings"
 	"time"
 
 	"github.com/zeta-chain/zetacore/zetaclient/bitcoin"
@@ -17,7 +16,6 @@ import (
 	"github.com/pkg/errors"
 	prom "github.com/prometheus/client_golang/prometheus"
 	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
 	"github.com/zeta-chain/zetacore/common"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	"github.com/zeta-chain/zetacore/zetaclient/config"
@@ -407,22 +405,4 @@ func (co *CoreObserver) getTargetChainOb(chainID int64) (interfaces.ChainClient,
 		return nil, fmt.Errorf("chain client not found for chainID %d", chainID)
 	}
 	return chainOb, nil
-}
-
-// HandleBroadcastError returns whether to retry in a few seconds, and whether to report via AddTxHashToOutTxTracker
-func HandleBroadcastError(err error, nonce, toChain, outTxHash string) (bool, bool) {
-	if strings.Contains(err.Error(), "nonce too low") {
-		log.Warn().Err(err).Msgf("nonce too low! this might be a unnecessary key-sign. increase re-try interval and awaits outTx confirmation")
-		return false, false
-	}
-	if strings.Contains(err.Error(), "replacement transaction underpriced") {
-		log.Warn().Err(err).Msgf("Broadcast replacement: nonce %s chain %s outTxHash %s", nonce, toChain, outTxHash)
-		return false, false
-	} else if strings.Contains(err.Error(), "already known") { // this is error code from QuickNode
-		log.Warn().Err(err).Msgf("Broadcast duplicates: nonce %s chain %s outTxHash %s", nonce, toChain, outTxHash)
-		return false, true // report to tracker, because there's possibilities a successful broadcast gets this error code
-	}
-
-	log.Error().Err(err).Msgf("Broadcast error: nonce %s chain %s outTxHash %s; retrying...", nonce, toChain, outTxHash)
-	return true, false
 }
